Split widget: diff token
